# Conflicts:
#	src/main/resources/templates/air/search-list-oneway.html

+++ src/main/java/go/travel/dnh/domain/pay/PayDTO.java
@@ -14,7 +14,7 @@ import java.util.Date;
 public class PayDTO {
 
     private String pno;
-    private Integer rno;
+    private Long rno;
     private Integer mno;
     private Integer pay_tot_price;
     private Date pay_date;
